+++ lib/rubocop_todo_corrector/rubocop_todo_section_parser.rb
@@ -28,9 +28,9 @@ module RubocopTodoCorrector
 
     # @return [Boolean]
     def auto_correctable?
-      @content.include?('# Cop supports --auto-correct.') ||
-        @content.include?('# Cop supports safe auto-correction.') ||
-        @content.include?('# Cop supports unsafe auto-correction.')
+      @content.include?('# Cop supports --auto-correct') ||
+        @content.include?('# This cop supports safe auto-correction') ||
+        @content.include?('# This cop supports unsafe auto-correction')
     end
 
     # @return [String, nil]