+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Abort if no cop was picked on `pick` command.
 
+### Fixed
+
+- Fix bug that `pick` cannot find any cop on rubocop >= 1.26.
+
 ## 0.1.0
 
 ### Added